refactor: AI chat panel layout shell with slim empty state and single-row composer footer
