+++ TablePro/Core/Services/Infrastructure/MainWindowToolbar.swift
@@ -10,7 +10,7 @@
 //  items to — NSToolbar must be constructed directly on NSWindow.
 //
 //  Each item's content is still authored in SwiftUI (`NSHostingView(rootView:)`)
-//  so existing subviews (ConnectionStatusView, SafeModeBadgeView, popovers,
+//  so existing subviews (ConnectionStatusView, SafeModeBadgeView, popovers,x
 //  etc.) are reused verbatim.
 //
 
@@ -39,11 +39,6 @@ internal final class MainWindowToolbar: NSObject, NSToolbarDelegate {
     private var sidebarButtons: [NSButton] = []
     private var sidebarObservationTask: Task<Void, Never>?
     private var splitViewObserver: NSObjectProtocol?
-
-    /// Strong refs to AI-specific toolbar items so we can flip `isHidden` on them
-    /// when `RightPanelState.activeTab` changes. Built lazily in `itemForItemIdentifier`.
-    internal var aiChatToolbarItems: [NSToolbarItem.Identifier: NSToolbarItem] = [:]
-    private var rightPanelObservationTask: Task<Void, Never>?
 
     internal init(coordinator: MainContentCoordinator) {
         self.coordinator = coordinator
@@ -75,15 +70,12 @@ internal final class MainWindowToolbar: NSObject, NSToolbarDelegate {
     func invalidate() {
         sidebarObservationTask?.cancel()
         sidebarObservationTask = nil
-        rightPanelObservationTask?.cancel()
-        rightPanelObservationTask = nil
         if let observer = splitViewObserver {
             NotificationCenter.default.removeObserver(observer)
             splitViewObserver = nil
         }
         sidebarButtons = []
         hostingControllers.removeAll()
-        aiChatToolbarItems.removeAll()
         coordinator = nil
     }
 
@@ -122,11 +114,6 @@ internal final class MainWindowToolbar: NSObject, NSToolbarDelegate {
             Self.newTab,
             Self.filters,
             Self.previewSQL,
-            Self.aiModePicker,
-            Self.aiModelPicker,
-            Self.aiNewConversation,
-            Self.aiHistory,
-            Self.panelTabPicker,
             Self.inspector,
         ]
     }
@@ -224,63 +211,8 @@ internal final class MainWindowToolbar: NSObject, NSToolbarDelegate {
                                    ExportToolbarButton(coordinator: coordinator)
                                    ImportToolbarButton(coordinator: coordinator)
                                })
-        case Self.panelTabPicker:
-            return registerAIItem(makePanelTabPickerItem(coordinator: coordinator), coordinator: coordinator)
-        case Self.aiModePicker:
-            return registerAIItem(makeAIModePickerItem(coordinator: coordinator), coordinator: coordinator)
-        case Self.aiModelPicker:
-            return registerAIItem(makeAIModelPickerItem(coordinator: coordinator), coordinator: coordinator)
-        case Self.aiNewConversation:
-            return registerAIItem(makeAINewConversationItem(coordinator: coordinator), coordinator: coordinator)
-        case Self.aiHistory:
-            return registerAIItem(makeAIHistoryItem(coordinator: coordinator), coordinator: coordinator)
         default:
             return nil
-        }
-    }
-
-    /// Records an AI-specific toolbar item so `syncAIChatItemVisibility`
-    /// can flip its `isHidden` flag, then kicks off observation if needed.
-    private func registerAIItem(
-        _ item: NSToolbarItem,
-        coordinator: MainContentCoordinator
-    ) -> NSToolbarItem {
-        aiChatToolbarItems[item.itemIdentifier] = item
-        syncAIChatItemVisibility(coordinator: coordinator)
-        startRightPanelObservation(coordinator: coordinator)
-        return item
-    }
-
-    /// Hide AI-only toolbar items unless the inspector's active tab is AI Chat.
-    /// The panel-tab picker stays visible whenever AI is enabled because it's
-    /// how the user gets to the AI tab in the first place.
-    fileprivate func syncAIChatItemVisibility(coordinator: MainContentCoordinator) {
-        let isAIEnabled = AppSettingsManager.shared.ai.enabled
-        let isAITabActive = isAIEnabled && coordinator.rightPanelState?.activeTab == .aiChat
-        for identifier in Self.aiChatItemIdentifiers {
-            aiChatToolbarItems[identifier]?.isHidden = !isAITabActive
-        }
-        aiChatToolbarItems[Self.panelTabPicker]?.isHidden = !isAIEnabled
-    }
-
-    fileprivate func startRightPanelObservation(coordinator: MainContentCoordinator) {
-        guard rightPanelObservationTask == nil else { return }
-        rightPanelObservationTask = Task { [weak self, weak coordinator] in
-            while !Task.isCancelled {
-                guard let coordinator else { return }
-                await withCheckedContinuation { continuation in
-                    withObservationTracking {
-                        _ = coordinator.rightPanelState?.activeTab
-                        _ = AppSettingsManager.shared.ai.enabled
-                    } onChange: {
-                        continuation.resume()
-                    }
-                }
-                guard !Task.isCancelled, let self else { return }
-                await MainActor.run {
-                    self.syncAIChatItemVisibility(coordinator: coordinator)
-                }
-            }
         }
     }
