+++ TablePro/Views/RightSidebar/UnifiedRightPanelView.swift
@@ -2,9 +2,6 @@
 //  UnifiedRightPanelView.swift
 //  TablePro
 //
-//  Unified right panel combining Details and AI Chat into a single
-//  segmented panel, reducing clutter and preserving AI conversation state.
-//
 
 import SwiftUI
 
@@ -12,10 +9,49 @@ struct UnifiedRightPanelView: View {
     @Bindable var state: RightPanelState
     let connection: DatabaseConnection
 
-    private var ctx: InspectorContext { state.inspectorContext }
+    private let settingsManager = AppSettingsManager.shared
+
+    var body: some View {
+        Group {
+            if settingsManager.ai.enabled {
+                splitContent
+            } else {
+                detailsView
+            }
+        }
+        .onChange(of: settingsManager.ai.enabled) {
+            if !settingsManager.ai.enabled {
+                state.activeTab = .details
+            }
+        }
+    }
+
+    private var splitContent: some View {
+        VStack(spacing: 0) {
+            tabPicker
+            Divider()
+            switch state.activeTab {
+            case .details: detailsView
+            case .aiChat:  aiChatView
+            }
+        }
+    }
+
+    private var tabPicker: some View {
+        Picker("", selection: $state.activeTab) {
+            ForEach(RightPanelTab.allCases, id: \.self) { tab in
+                Text(tab.localizedTitle).tag(tab)
+            }
+        }
+        .pickerStyle(.segmented)
+        .labelsHidden()
+        .padding(.horizontal, 12)
+        .padding(.vertical, 6)
+    }
 
     private var detailsView: some View {
-        RightSidebarView(
+        let ctx = state.inspectorContext
+        return RightSidebarView(
             tableName: ctx.tableName,
             tableMetadata: ctx.tableMetadata,
             selectedRowData: ctx.selectedRowData,
@@ -26,28 +62,13 @@ struct UnifiedRightPanelView: View {
         )
     }
 
-    var body: some View {
-        Group {
-            if AppSettingsManager.shared.ai.enabled {
-                switch state.activeTab {
-                case .details:
-                    detailsView
-                case .aiChat:
-                    AIChatPanelView(
-                        connection: connection,
-                        currentQuery: ctx.currentQuery,
-                        queryResults: ctx.queryResults,
-                        viewModel: state.aiViewModel
-                    )
-                }
-            } else {
-                detailsView
-            }
-        }
-        .onChange(of: AppSettingsManager.shared.ai.enabled) {
-            if !AppSettingsManager.shared.ai.enabled {
-                state.activeTab = .details
-            }
-        }
+    private var aiChatView: some View {
+        let ctx = state.inspectorContext
+        return AIChatPanelView(
+            connection: connection,
+            currentQuery: ctx.currentQuery,
+            queryResults: ctx.queryResults,
+            viewModel: state.aiViewModel
+        )
     }
 }
