+++ docs/features/ai-assistant.mdx
@@ -43,7 +43,7 @@ Add Copilot like any other provider. The detail sheet runs GitHub's device-flow 
 
 ## Chat
 
-Press `Cmd+Shift+L`, click the AI toolbar button, or use **View** > **Toggle AI Chat**.
+Press `Cmd+Shift+L`, click the inspector toggle and pick **AI Chat**, or use **View** > **Toggle AI Chat**. The right inspector has a Details / AI Chat segmented picker at the top.
 
 <Frame caption="AI chat panel">
   <img
@@ -86,7 +86,7 @@ Type a question and press Return. The AI receives your database type, schema, an
   />
 </Frame>
 
-Conversations auto-save and auto-title from your first message. Browse, clear, or start a new one from the panel header.
+Conversations auto-save and auto-title from your first message. Browse, clear, or start a new one from the clock and pencil icons in the composer footer.
 
 <Frame caption="Conversation history">
   <img
